[issue-118] Transfer a REGO checker to the separate package
